+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/model/WorkloadUtilsTest.java
@@ -19,6 +19,9 @@ import io.fabric8.kubernetes.api.model.OwnerReference;
 import io.fabric8.kubernetes.api.model.OwnerReferenceBuilder;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
+import io.fabric8.kubernetes.api.model.PodDNSConfig;
+import io.fabric8.kubernetes.api.model.PodDNSConfigBuilder;
+import io.fabric8.kubernetes.api.model.PodDNSConfigOptionBuilder;
 import io.fabric8.kubernetes.api.model.PodSecurityContext;
 import io.fabric8.kubernetes.api.model.PodSecurityContextBuilder;
 import io.fabric8.kubernetes.api.model.PodTemplateSpec;
@@ -30,6 +33,7 @@ import io.fabric8.kubernetes.api.model.TopologySpreadConstraintBuilder;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.DeploymentStrategy;
 import io.strimzi.api.kafka.model.common.template.DeploymentTemplateBuilder;
+import io.strimzi.api.kafka.model.common.template.DnsPolicy;
 import io.strimzi.api.kafka.model.common.template.PodTemplate;
 import io.strimzi.api.kafka.model.common.template.PodTemplateBuilder;
 import io.strimzi.api.kafka.model.common.template.ResourceTemplateBuilder;
@@ -120,6 +124,20 @@ public class WorkloadUtilsTest {
     private static final HostAlias DEFAULT_HOST_ALIAS = new HostAliasBuilder()
             .withIp("127.0.0.1")
             .withHostnames("home")
+            .build();
+    private static final DnsPolicy DEFAULT_DNS_POLICY = DnsPolicy.NONE;
+    private static final PodDNSConfig DEFAULT_DNS_CONFIG = new PodDNSConfigBuilder()
+            .withNameservers("192.0.2.1")
+            .withSearches("ns1.svc.cluster-domain.example", "my.dns.search.suffix")
+            .withOptions(
+                new PodDNSConfigOptionBuilder()
+                        .withName("ndots")
+                        .withValue("2")
+                        .build(), 
+                new PodDNSConfigOptionBuilder()
+                        .withName("edns0")
+                        .build()
+            )
             .build();
 
     //////////////////////////////////////////////////
@@ -452,6 +470,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -502,6 +522,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -552,6 +574,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -574,6 +598,8 @@ public class WorkloadUtilsTest {
                         .withImagePullSecrets(List.of(new LocalObjectReference("some-other-pull-secret")))
                         .withPriorityClassName("my-priority-class")
                         .withHostAliases(DEFAULT_HOST_ALIAS)
+                        .withDnsPolicy(DEFAULT_DNS_POLICY)
+                        .withDnsConfig(DEFAULT_DNS_CONFIG)
                         .withTolerations(DEFAULT_TOLERATION)
                         .withTerminationGracePeriodSeconds(15)
                         .withSecurityContext(new PodSecurityContextBuilder().withRunAsUser(0L).build()) // => should be ignored
@@ -598,7 +624,7 @@ public class WorkloadUtilsTest {
                 .withStrimziPodName(NAME + "-0")
                 .withAdditionalLabels(Map.of("statefulset.kubernetes.io/pod-name", "my-workload-0", "default-label", "default-value", "label-3", "value-3", "label-4", "value-4"))
                 .toMap()));
-        assertThat(pod.getMetadata().getAnnotations(), is(Map.of("extra", "annotations", "anno-1", "value-1", "anno-2", "value-2", PodRevision.STRIMZI_REVISION_ANNOTATION, "4c2e5618")));
+        assertThat(pod.getMetadata().getAnnotations(), is(Map.of("extra", "annotations", "anno-1", "value-1", "anno-2", "value-2", PodRevision.STRIMZI_REVISION_ANNOTATION, "d3ffc657")));
 
         assertThat(pod.getSpec().getRestartPolicy(), is("Always"));
         assertThat(pod.getSpec().getHostname(), is(NAME + "-0"));
@@ -617,6 +643,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is("my-priority-class"));
         assertThat(pod.getSpec().getSchedulerName(), is("my-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(List.of(DEFAULT_HOST_ALIAS)));
+        assertThat(pod.getSpec().getDnsPolicy(), is(DEFAULT_DNS_POLICY.toValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(DEFAULT_DNS_CONFIG));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(List.of(DEFAULT_TOPOLOGY_SPREAD_CONSTRAINT)));
     }
 
@@ -657,6 +685,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -694,6 +724,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -731,6 +763,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -749,6 +783,8 @@ public class WorkloadUtilsTest {
                         .withImagePullSecrets(List.of(new LocalObjectReference("some-other-pull-secret")))
                         .withPriorityClassName("my-priority-class")
                         .withHostAliases(DEFAULT_HOST_ALIAS)
+                        .withDnsPolicy(DEFAULT_DNS_POLICY)
+                        .withDnsConfig(DEFAULT_DNS_CONFIG)
                         .withTolerations(DEFAULT_TOLERATION)
                         .withTerminationGracePeriodSeconds(15)
                         .withSecurityContext(new PodSecurityContextBuilder().withRunAsUser(0L).build()) // => should be ignored
@@ -783,6 +819,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is("my-priority-class"));
         assertThat(pod.getSpec().getSchedulerName(), is("my-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(List.of(DEFAULT_HOST_ALIAS)));
+        assertThat(pod.getSpec().getDnsPolicy(), is(DEFAULT_DNS_POLICY.toValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(DEFAULT_DNS_CONFIG));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(List.of(DEFAULT_TOPOLOGY_SPREAD_CONSTRAINT)));
     }
 
@@ -828,6 +866,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -870,6 +910,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -912,6 +954,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is(nullValue()));
         assertThat(pod.getSpec().getSchedulerName(), is("default-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsPolicy(), is(nullValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(nullValue()));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(nullValue()));
     }
 
@@ -932,6 +976,8 @@ public class WorkloadUtilsTest {
                         .withImagePullSecrets(List.of(new LocalObjectReference("some-other-pull-secret")))
                         .withPriorityClassName("my-priority-class")
                         .withHostAliases(DEFAULT_HOST_ALIAS)
+                        .withDnsPolicy(DEFAULT_DNS_POLICY)
+                        .withDnsConfig(DEFAULT_DNS_CONFIG)
                         .withTolerations(DEFAULT_TOLERATION)
                         .withTerminationGracePeriodSeconds(15)
                         .withSecurityContext(new PodSecurityContextBuilder().withRunAsUser(0L).build()) // => should be ignored
@@ -969,6 +1015,8 @@ public class WorkloadUtilsTest {
         assertThat(pod.getSpec().getPriorityClassName(), is("my-priority-class"));
         assertThat(pod.getSpec().getSchedulerName(), is("my-scheduler"));
         assertThat(pod.getSpec().getHostAliases(), is(List.of(DEFAULT_HOST_ALIAS)));
+        assertThat(pod.getSpec().getDnsPolicy(), is(DEFAULT_DNS_POLICY.toValue()));
+        assertThat(pod.getSpec().getDnsConfig(), is(DEFAULT_DNS_CONFIG));
         assertThat(pod.getSpec().getTopologySpreadConstraints(), is(List.of(DEFAULT_TOPOLOGY_SPREAD_CONSTRAINT)));
     }