+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support for ZooKeeper-based Apache Kafka clusters and for KRaft migration has been removed
 * Support for MirrorMaker 1 has been removed
+* Added support to configure `dnsPolicy` and `dnsConfig` using the `template` sections.
 
 ### Major changes, deprecations and removals